refactor(node editor): moved out nodes specific view code from main view to nodeEditor component
